pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pmg"
version = "0.1.0"
description = "Investigating Methods for Polymicrogyria Classification using Deep Learning."
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "invoke>=2.2.1",
    # --- Deep Learning ---
    "torch",         # The primary deep learning framework.
    "torchvision",   # Provides models, datasets, and transforms for computer vision.
    # --- Data Science & Numeric Computing ---
    "numpy",         # Fundamental for numerical operations.
    "pandas",        # For handling metadata and experiment results (e.g., from CSVs).
    # --- Medical Imaging & Data Visualization ---
    "nibabel",       # The standard library for reading/writing neuroimaging files (like NIfTI).
    "scikit-image",  # For general image processing and feature extraction.
    "matplotlib",    # For creating plots and visualizations.
    "seaborn",       # For creating more attractive statistical plots.
    # --- Experimentation & Utilities ---
    "tensorboard",   # For visualizing model training progress.
    "jupyterlab",    # For interactive exploration and prototyping in notebooks.
    "tqdm",          # For displaying smart progress bars in loops.
    "hydra-core>=1.3.2",
    "opencv-python>=4.13.0.92",
    "omegaconf>=2.3.0",
    "natsort>=8.4.0",
    "scikit-learn>=1.8.0",
]

[project.scripts]
preprocess = "src.cli.preprocess:preprocess_cli"

# == Optional/Development Dependencies ==
# These are tools for code quality, testing, and development.
# You can install them with `uv pip install -e .[dev]`
[project.optional-dependencies]
dev = [
    "ruff",          # High-performance linter.
    "mypy",          # Static type checker.
    "pytest",        # Framework for writing and running tests.
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "examples*"]


